Add C_n and balance_n time series to -lay_massfile output.  Needs work.
Recall principle that
  M_n = M_{n-1} + C_n - R_n
for time-semi-discretized problem.  This code computes C_n and then computes
  balance_n = M_n - M_{n-1} - C_n + R_n
which should come out to zero.

But in space-discretized case we don't (as seen when running this version)
get balance_n = 0.  The reason is that
  Q_{j+1/2}^n
can be nonzero even if one of u_j^n or u_{j+1}^n is zero.  We want to
add that property, i.e.
  zero flux at j+1/2 if one of u_j^n or u_{j+1}^n is zero
Then return to testing this code.